halatuju_api/requirements.txt
# Django and REST
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# JWT verification (for Supabase Auth)
PyJWT>=2.8,<3.0

# Data processing (hybrid engine approach)
pandas>=2.0,<3.0
numpy>=1.24,<2.0

# Production server
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0

# Observability
sentry-sdk>=1.39,<2.0

# AI (for reports)
google-generativeai>=0.3,<1.0

# Development
python-dotenv>=1.0,<2.0
